+++ src/components/Loader/Loader.tsx
@@ -1,0 +1,9 @@
+import styles from './Loader.module.scss';
+
+export function Loader() {
+  return (
+    <div className={styles.loaderContainer}>
+      <div className={styles.loader}></div>
+    </div>
+  );
+}
